+++ tests/unit/documentation-helpers.spec.ts
@@ -62,30 +62,36 @@ describe("Documentation helpers", () => {
       $ref: `#/components/schemas/${name}`,
     }),
   );
-  const requestCtx: OpenAPIContext = {
+  const requestCtx = {
     path: "/v1/user/:id",
     method: "get",
     isResponse: false,
     getRef: getRefMock,
     makeRef: makeRefMock,
     serializer: defaultSerializer,
-  };
-  const responseCtx: OpenAPIContext = {
+    next: (schema: z.ZodTypeAny) =>
+      walkSchema(schema, {
+        rules: depicters,
+        onEach,
+        onMissing,
+        ctx: requestCtx,
+      }),
+  } satisfies OpenAPIContext;
+  const responseCtx = {
     path: "/v1/user/:id",
     method: "get",
     isResponse: true,
     getRef: getRefMock,
     makeRef: makeRefMock,
     serializer: defaultSerializer,
-  };
-  const makeNext = (ctx: OpenAPIContext) => (schema: z.ZodTypeAny) =>
-    walkSchema({
-      schema,
-      rules: depicters,
-      ...ctx,
-      onEach,
-      onMissing,
-    });
+    next: (schema: z.ZodTypeAny) =>
+      walkSchema(schema, {
+        rules: depicters,
+        onEach,
+        onMissing,
+        ctx: responseCtx,
+      }),
+  } satisfies OpenAPIContext;
 
   beforeEach(() => {
     getRefMock.mockClear();
@@ -187,12 +193,11 @@ describe("Documentation helpers", () => {
       z
         .record(z.literal("a"), z.string())
         .and(z.record(z.string(), z.string())),
-    ])("should omit specified path params %#", (schema) => {
-      const depicted = walkSchema({
-        schema,
-        ...requestCtx,
-        onEach,
+    ])("should omit specified params %#", (schema) => {
+      const depicted = walkSchema(schema, {
+        ctx: requestCtx,
         rules: depicters,
+        onEach,
         onMissing,
       });
       expect(excludeParamsFromDepiction(depicted, ["a"])).toMatchSnapshot();
@@ -221,24 +226,19 @@ describe("Documentation helpers", () => {
   describe("depictDefault()", () => {
     test("should set default property", () => {
       expect(
-        depictDefault({
-          schema: z.boolean().default(true),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+        depictDefault(z.boolean().default(true), requestCtx),
       ).toMatchSnapshot();
     });
     test("Feature #1706: should override the default value by a label from metadata", () => {
       expect(
-        depictDefault({
-          schema: z
+        depictDefault(
+          z
             .string()
             .datetime()
             .default(() => new Date().toISOString())
             .label("Today"),
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
+          responseCtx,
+        ),
       ).toMatchSnapshot();
     });
   });
@@ -246,44 +246,24 @@ describe("Documentation helpers", () => {
   describe("depictCatch()", () => {
     test("should pass next depicter", () => {
       expect(
-        depictCatch({
-          schema: z.boolean().catch(true),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+        depictCatch(z.boolean().catch(true), requestCtx),
       ).toMatchSnapshot();
     });
   });
 
   describe("depictAny()", () => {
     test("should set format:any", () => {
-      expect(
-        depictAny({
-          schema: z.any(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictAny(z.any(), requestCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictUpload()", () => {
     test("should set format:binary and type:string", () => {
-      expect(
-        depictUpload({
-          schema: ez.upload(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictUpload(ez.upload(), requestCtx)).toMatchSnapshot();
     });
     test("should throw when using in response", () => {
       try {
-        depictUpload({
-          schema: ez.upload(),
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        });
+        depictUpload(ez.upload(), responseCtx);
         expect.fail("Should not be here");
       } catch (e) {
         expect(e).toBeInstanceOf(DocumentationError);
@@ -300,24 +280,14 @@ describe("Documentation helpers", () => {
       ez.file("string"),
       ez.file("buffer"),
     ])("should set type:string and format accordingly %#", (schema) => {
-      expect(
-        depictFile({
-          schema,
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictFile(schema, responseCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictUnion()", () => {
     test("should wrap next depicters into oneOf property", () => {
       expect(
-        depictUnion({
-          schema: z.string().or(z.number()),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+        depictUnion(z.string().or(z.number()), requestCtx),
       ).toMatchSnapshot();
     });
   });
@@ -325,17 +295,16 @@ describe("Documentation helpers", () => {
   describe("depictDiscriminatedUnion()", () => {
     test("should wrap next depicters in oneOf prop and set discriminator prop", () => {
       expect(
-        depictDiscriminatedUnion({
-          schema: z.discriminatedUnion("status", [
+        depictDiscriminatedUnion(
+          z.discriminatedUnion("status", [
             z.object({ status: z.literal("success"), data: z.any() }),
             z.object({
               status: z.literal("error"),
               error: z.object({ message: z.string() }),
             }),
           ]),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+          requestCtx,
+        ),
       ).toMatchSnapshot();
     });
   });
@@ -343,20 +312,17 @@ describe("Documentation helpers", () => {
   describe("depictIntersection()", () => {
     test("should flatten two object schemas", () => {
       expect(
-        depictIntersection({
-          schema: z
-            .object({ one: z.number() })
-            .and(z.object({ two: z.number() })),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+        depictIntersection(
+          z.object({ one: z.number() }).and(z.object({ two: z.number() })),
+          requestCtx,
+        ),
       ).toMatchSnapshot();
     });
 
     test("should merge examples deeply", () => {
       expect(
-        depictIntersection({
-          schema: z
+        depictIntersection(
+          z
             .object({ test: z.object({ a: z.number() }) })
             .example({ test: { a: 123 } })
             .and(
@@ -364,35 +330,32 @@ describe("Documentation helpers", () => {
                 .object({ test: z.object({ b: z.number() }) })
                 .example({ test: { b: 456 } }),
             ),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+          requestCtx,
+        ),
       ).toMatchSnapshot();
     });
 
     test("should flatten three object schemas with examples", () => {
       expect(
-        depictIntersection({
-          schema: z
+        depictIntersection(
+          z
             .object({ one: z.number() })
             .example({ one: 123 })
             .and(z.object({ two: z.number() }).example({ two: 456 }))
             .and(z.object({ three: z.number() }).example({ three: 789 })),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+          requestCtx,
+        ),
       ).toMatchSnapshot();
     });
 
     test("should maintain uniqueness in the array of required props", () => {
       expect(
-        depictIntersection({
-          schema: z
+        depictIntersection(
+          z
             .record(z.literal("test"), z.number())
             .and(z.object({ test: z.literal(5) })),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+          requestCtx,
+        ),
       ).toMatchSnapshot();
     });
 
@@ -400,42 +363,24 @@ describe("Documentation helpers", () => {
       z.record(z.string(), z.number()).and(z.object({ test: z.number() })), // has additionalProperties
       z.number().and(z.literal(5)), // not objects
     ])("should fall back to allOf in other cases %#", (schema) => {
-      expect(
-        depictIntersection({
-          schema,
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictIntersection(schema, requestCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictOptional()", () => {
-    test.each<OpenAPIContext>([requestCtx, responseCtx])(
+    test.each([requestCtx, responseCtx])(
       "should pass the next depicter %#",
       (ctx) => {
-        expect(
-          depictOptional({
-            schema: z.string().optional(),
-            ...ctx,
-            next: makeNext(ctx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictOptional(z.string().optional(), ctx)).toMatchSnapshot();
       },
     );
   });
 
   describe("depictNullable()", () => {
-    test.each<OpenAPIContext>([requestCtx, responseCtx])(
+    test.each([requestCtx, responseCtx])(
       "should add null to the type %#",
       (ctx) => {
-        expect(
-          depictNullable({
-            schema: z.string().nullable(),
-            ...ctx,
-            next: makeNext(ctx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictNullable(z.string().nullable(), ctx)).toMatchSnapshot();
       },
     );
 
@@ -444,13 +389,7 @@ describe("Documentation helpers", () => {
       z.null().nullable(),
       z.string().nullable().nullable(),
     ])("should only add null type once %#", (schema) => {
-      expect(
-        depictNullable({
-          schema,
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictNullable(schema, requestCtx)).toMatchSnapshot();
     });
   });
 
@@ -462,13 +401,7 @@ describe("Documentation helpers", () => {
     test.each([z.enum(["one", "two"]), z.nativeEnum(Test)])(
       "should set type and enum properties",
       (schema) => {
-        expect(
-          depictEnum({
-            schema,
-            ...requestCtx,
-            next: makeNext(requestCtx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictEnum(schema, requestCtx)).toMatchSnapshot();
       },
     );
   });
@@ -477,19 +410,13 @@ describe("Documentation helpers", () => {
     test.each(["testng", null, BigInt(123), Symbol("test")])(
       "should set type and involve const property %#",
       (value) => {
-        expect(
-          depictLiteral({
-            schema: z.literal(value),
-            ...requestCtx,
-            next: makeNext(requestCtx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictLiteral(z.literal(value), requestCtx)).toMatchSnapshot();
       },
     );
   });
 
   describe("depictObject()", () => {
-    test.each<{ ctx: OpenAPIContext; shape: z.ZodRawShape }>([
+    test.each([
       { ctx: requestCtx, shape: { a: z.number(), b: z.string() } },
       { ctx: responseCtx, shape: { a: z.number(), b: z.string() } },
       {
@@ -504,13 +431,7 @@ describe("Documentation helpers", () => {
     ])(
       "should type:object, properties and required props %#",
       ({ shape, ctx }) => {
-        expect(
-          depictObject({
-            schema: z.object(shape),
-            ...ctx,
-            next: makeNext(ctx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictObject(z.object(shape), ctx)).toMatchSnapshot();
       },
     );
 
@@ -520,49 +441,25 @@ describe("Documentation helpers", () => {
         b: z.coerce.string(),
         c: z.coerce.string().optional(),
       });
-      expect(
-        depictObject({
-          schema,
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictObject(schema, responseCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictNull()", () => {
     test("should give type:null", () => {
-      expect(
-        depictNull({
-          schema: z.null(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictNull(z.null(), requestCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictBoolean()", () => {
     test("should set type:boolean", () => {
-      expect(
-        depictBoolean({
-          schema: z.boolean(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictBoolean(z.boolean(), requestCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictBigInt()", () => {
     test("should set type:integer and format:bigint", () => {
-      expect(
-        depictBigInt({
-          schema: z.bigint(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictBigInt(z.bigint(), requestCtx)).toMatchSnapshot();
     });
   });
 
@@ -577,68 +474,39 @@ describe("Documentation helpers", () => {
     ])(
       "should set properties+required or additionalProperties props %#",
       (schema) => {
-        expect(
-          depictRecord({
-            schema,
-            ...requestCtx,
-            next: makeNext(requestCtx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictRecord(schema, requestCtx)).toMatchSnapshot();
       },
     );
   });
 
   describe("depictArray()", () => {
     test("should set type:array and pass items depiction", () => {
-      expect(
-        depictArray({
-          schema: z.array(z.boolean()),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictArray(z.array(z.boolean()), requestCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictTuple()", () => {
     test("should utilize prefixItems and set items:not:{}", () => {
       expect(
-        depictTuple({
-          schema: z.tuple([z.boolean(), z.string(), z.literal("test")]),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+        depictTuple(
+          z.tuple([z.boolean(), z.string(), z.literal("test")]),
+          requestCtx,
+        ),
       ).toMatchSnapshot();
     });
     test("should depict rest as items when defined", () => {
       expect(
-        depictTuple({
-          schema: z.tuple([z.boolean()]).rest(z.string()),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+        depictTuple(z.tuple([z.boolean()]).rest(z.string()), requestCtx),
       ).toMatchSnapshot();
     });
     test("should depict empty tuples as is", () => {
-      expect(
-        depictTuple({
-          schema: z.tuple([]),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictTuple(z.tuple([]), requestCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictString()", () => {
     test("should set type:string", () => {
-      expect(
-        depictString({
-          schema: z.string(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictString(z.string(), requestCtx)).toMatchSnapshot();
     });
 
     test.each([
@@ -650,13 +518,7 @@ describe("Documentation helpers", () => {
       z.string().datetime({ offset: true }),
       z.string().regex(/^\d+.\d+.\d+$/),
     ])("should set format, pattern and min/maxLength props %#", (schema) => {
-      expect(
-        depictString({
-          schema,
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictString(schema, requestCtx)).toMatchSnapshot();
     });
   });
 
@@ -664,13 +526,7 @@ describe("Documentation helpers", () => {
     test.each([z.number(), z.number().int().min(10).max(20)])(
       "should type:number, min/max, format and exclusiveness props",
       (schema) => {
-        expect(
-          depictNumber({
-            schema,
-            ...requestCtx,
-            next: makeNext(requestCtx),
-          }),
-        ).toMatchSnapshot();
+        expect(depictNumber(schema, requestCtx)).toMatchSnapshot();
       },
     );
   });
@@ -678,24 +534,19 @@ describe("Documentation helpers", () => {
   describe("depictObjectProperties()", () => {
     test("should wrap next depicters in a shape of object", () => {
       expect(
-        depictObjectProperties({
-          schema: z.object({
+        depictObjectProperties(
+          z.object({
             one: z.string(),
             two: z.boolean(),
           }),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
+          requestCtx.next,
+        ),
       ).toMatchSnapshot();
     });
   });
 
   describe("depictEffect()", () => {
-    test.each<{
-      ctx: OpenAPIContext;
-      schema: z.ZodEffects<any>;
-      expected: string;
-    }>([
+    test.each([
       {
         schema: z.string().transform((v) => parseInt(v, 10)),
         ctx: responseCtx,
@@ -719,40 +570,24 @@ describe("Documentation helpers", () => {
         expected: "object (refinement)",
       },
     ])("should depict as $expected", ({ schema, ctx }) => {
-      expect(
-        depictEffect({
-          schema,
-          ...ctx,
-          next: makeNext(ctx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictEffect(schema, ctx)).toMatchSnapshot();
     });
 
     test.each([
       z.number().transform((num) => () => num),
       z.number().transform(() => assert.fail("this should be handled")),
     ])("should handle edge cases", (schema) => {
-      expect(
-        depictEffect({
-          schema,
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictEffect(schema, responseCtx)).toMatchSnapshot();
     });
   });
 
   describe("depictPipeline", () => {
-    test.each<{ ctx: OpenAPIContext; expected: string }>([
+    test.each([
       { ctx: responseCtx, expected: "boolean (out)" },
       { ctx: requestCtx, expected: "string (in)" },
     ])("should depict as $expected", ({ ctx }) => {
       expect(
-        depictPipeline({
-          schema: z.string().pipe(z.coerce.boolean()),
-          ...ctx,
-          next: makeNext(ctx),
-        }),
+        depictPipeline(z.string().pipe(z.coerce.boolean()), ctx),
       ).toMatchSnapshot();
     });
   });
@@ -886,21 +721,11 @@ describe("Documentation helpers", () => {
 
   describe("depictDateIn", () => {
     test("should set type:string, pattern and format", () => {
-      expect(
-        depictDateIn({
-          schema: ez.dateIn(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictDateIn(ez.dateIn(), requestCtx)).toMatchSnapshot();
     });
     test("should throw when ZodDateIn in response", () => {
       try {
-        depictDateIn({
-          schema: ez.dateIn(),
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        });
+        depictDateIn(ez.dateIn(), responseCtx);
         expect.fail("should not be here");
       } catch (e) {
         expect(e).toBeInstanceOf(DocumentationError);
@@ -911,21 +736,11 @@ describe("Documentation helpers", () => {
 
   describe("depictDateOut", () => {
     test("should set type:string, description and format", () => {
-      expect(
-        depictDateOut({
-          schema: ez.dateOut(),
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictDateOut(ez.dateOut(), responseCtx)).toMatchSnapshot();
     });
     test("should throw when ZodDateOut in request", () => {
       try {
-        depictDateOut({
-          schema: ez.dateOut(),
-          ...requestCtx,
-          next: makeNext(requestCtx),
-        });
+        depictDateOut(ez.dateOut(), requestCtx);
         expect.fail("should not be here");
       } catch (e) {
         expect(e).toBeInstanceOf(DocumentationError);
@@ -935,15 +750,11 @@ describe("Documentation helpers", () => {
   });
 
   describe("depictDate", () => {
-    test.each<OpenAPIContext>([responseCtx, requestCtx])(
+    test.each([responseCtx, requestCtx])(
       "should throw clear error %#",
       (ctx) => {
         try {
-          depictDate({
-            schema: z.date(),
-            ...ctx,
-            next: makeNext(ctx),
-          });
+          depictDate(z.date(), ctx);
           expect.fail("should not be here");
         } catch (e) {
           expect(e).toBeInstanceOf(DocumentationError);
@@ -956,11 +767,7 @@ describe("Documentation helpers", () => {
   describe("depictBranded", () => {
     test("should pass the next depicter", () => {
       expect(
-        depictBranded({
-          schema: z.string().min(2).brand<"Test">(),
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
+        depictBranded(z.string().min(2).brand("Test"), responseCtx),
       ).toMatchSnapshot();
     });
   });
@@ -968,11 +775,7 @@ describe("Documentation helpers", () => {
   describe("depictReadonly", () => {
     test("should pass the next depicter", () => {
       expect(
-        depictReadonly({
-          schema: z.string().readonly(),
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
+        depictReadonly(z.string().readonly(), responseCtx),
       ).toMatchSnapshot();
     });
   });
@@ -1008,13 +811,7 @@ describe("Documentation helpers", () => {
           }),
         );
       expect(getRefMock.mock.calls.length).toBe(0);
-      expect(
-        depictLazy({
-          schema,
-          ...responseCtx,
-          next: makeNext(responseCtx),
-        }),
-      ).toMatchSnapshot();
+      expect(depictLazy(schema, responseCtx)).toMatchSnapshot();
       expect(getRefMock).toHaveBeenCalledTimes(2);
       for (const call of getRefMock.mock.calls) {
         expect(call[0]).toBe(hash);